+++ Model/Notify.php
@@ -288,21 +288,22 @@ class Notify
                 break;
             case TransactionStatus::CAPTURE_REQUESTED:
             case TransactionStatus::CAPTURED:
-                // if operation ID exists matching magento2, check invoice related to this order
-                // then, if invoice does not exist ~> refuse notif
+                // If an operation ID exists and matches the Magento manual capture format,
+                // check if an invoice is already linked to this order.
+                // If no invoice is found, create a new one and assign the operation ID.
                 $operationId = $this->_transaction->getOperation()
                     ? $this->_transaction->getOperation()->getId()
                     : null;
                 if (
                     $operationId
-                        && preg_match("/-" . Operation::CAPTURE . "-manual-/", $operationId)
-                        && !$this->getInvoiceForTransactionId($this->_order, $operationId)
+                    && preg_match("/-" . Operation::CAPTURE . "-manual-/", $operationId)
                 ) {
-                    throw new WebApiException(
-                        __(sprintf('Invoice "%s" does not exist in database.', $operationId)),
-                        0,
-                        WebApiException::HTTP_BAD_REQUEST
-                    );
+                    $invoice = $this->getInvoiceForTransactionId($this->_order, $operationId);
+                    if (!$invoice) {
+                        $invoice = $this->_order->prepareInvoice()->register();
+                        $invoice->setTransactionId($operationId);
+                        $this->_order->addRelatedObject($invoice);
+                    }
                 }
 
                 // status : 118 - We check the 116 has been received before handling