+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## UNRELEASE
 
+- **Fix** : Enhanced payment method validation with smart button visibility control for PayPal and Apple Pay.
+
 ## 1.29.0
 
 - **Add** : Made the cancellation delay configurable per payment method.
 - **Add** : Use JS SDK for Oney, add Oney widget & include oney.io domain in CSP
 - **Fix** : Improved RestoreBasketObserver to handle missing order during early HiPay notifications.
-- **Fix** : Enhanced payment method validation with smart button visibility control for PayPal and Apple Pay.
 
 ## 1.28.0