+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASE
 
+## 1.29.1
+
 - **Fix** : Enhanced payment method validation with smart button visibility control for PayPal and Apple Pay.
 - **Fix** : Handle manual capture notifications from Magento when HiPay response arrives before invoice creation, by generating the invoice with the operationId to avoid blocked orders.
 - **Fix** : Enhanced One-Click functionality by adding a layer of validation for hosted field payment method